#649 Move background property builders to own file.
So they can easily be edited to all allow multiple values.
